git-svn-id: file:///mnt/data/Workspace/zie87/game_development/tmp/hare-game-engine/p/trunk@283 67360d68-41fe-4630-88ca-a5c15c9c3197

+++ plugins/gl_renderer/Win32/GLRenderWindow.cpp
@@ -153,12 +153,12 @@ void GLRenderWindow::create(const WindowParams& params)
 
 void GLRenderWindow::setProjection()
 {
-    int l = cameraPos.x;
+    int l = (int)cameraPos.x;
     int r = windowParams.width + (int)cameraPos.x;
     int b = windowParams.height + (int)cameraPos.y;
-    int t = cameraPos.y;
+    int t = (int)cameraPos.y;
 
-    RenderSystem::getSingletonPtr()->setProjection(l, r, b, t);
+    RenderSystem::getSingletonPtr()->setProjection((float)l, (float)r, (float)b, (float)t);
 }
 
 void GLRenderWindow::resize(uint32 w, uint32 h)
